Use explicit imports and move host.py only code from util.py.
