+++ mreg-cli/util.py
@@ -290,40 +290,6 @@ def clean_hostname(name: typing.AnyStr) -> str:
             return "{}.{}".format(name, conf['domain'])
     return name
 
-################################################################################
-#                                                                              #
-#   Hinfo utility                                                              #
-#                                                                              #
-################################################################################
-
-HinfoTuple = typing.Tuple[str, str]
-HinfoDict = typing.Dict[int, HinfoTuple]
-
-def hinfo_sanify(hid: str, hinfo: HinfoDict):
-    """Check if the requested hinfo is a valid one."""
-    try:
-        int(hid)
-    except ValueError:
-        cli_warning("hinfo {} is not a number".format(hid))
-    if len(hinfo) == 0:
-        cli_warning("Can not set hinfo, as no hinfo presets defined")
-    if not hid in hinfo:
-        cli_warning("Unknown hinfo preset {}".format(hid))
-
-def hinfo_dict() -> HinfoDict:
-    """
-    Return a dict with descriptions of available hinfo presets. The keys
-    are the hinfo ids.
-    """
-    path = "/hinfopresets/"
-    history.record_get(path)
-    hinfo_get = get(path)
-    hl = dict()
-    for hinfo in hinfo_get.json():
-        assert isinstance(hinfo, dict)
-        hl[str(hinfo["id"])] = (hinfo["cpu"], hinfo["os"])
-    return hl
-
 
 ################################################################################
 #                                                                              #
@@ -404,157 +370,6 @@ def string_to_int(value, error_tag):
 
 ################################################################################
 #                                                                              #
-#   Pretty printing                                                            #
-#                                                                              #
-################################################################################
-
-
-def print_host_name(name: str, padding: int = 14) -> None:
-    """Pretty print given name."""
-    if name is None:
-        return
-    assert isinstance(name, str)
-    print("{1:<{0}}{2}".format(padding, "Name:", name))
-
-
-def print_contact(contact: str, padding: int = 14) -> None:
-    """Pretty print given contact."""
-    if contact is None:
-        return
-    assert isinstance(contact, str)
-    print("{1:<{0}}{2}".format(padding, "Contact:", contact))
-
-
-def print_comment(comment: str, padding: int = 14) -> None:
-    """Pretty print given comment."""
-    if comment is None:
-        return
-    assert isinstance(comment, str)
-    print("{1:<{0}}{2}".format(padding, "Comment:", comment))
-
-
-def print_ipaddresses(ipaddresses: typing.Iterable[dict], padding: int = 14) -> None:
-    """Pretty print given ip addresses"""
-    if ipaddresses is None:
-        return
-    a_records = []
-    aaaa_records = []
-    len_ip = 0
-    for record in ipaddresses:
-        if is_valid_ipv4(record["ipaddress"]):
-            a_records.append(record)
-            if len(record["ipaddress"]) > len_ip:
-                len_ip = len(record["ipaddress"])
-        elif is_valid_ipv6(record["ipaddress"]):
-            aaaa_records.append(record)
-            if len(record["ipaddress"]) > len_ip:
-                len_ip = len(record["ipaddress"])
-    len_ip += 2
-    if a_records:
-        print("{1:<{0}}{2:<{3}}{4}".format(padding, "A_Records:", "IP", len_ip, "MAC"))
-        for record in a_records:
-            ip = record["ipaddress"]
-            mac = record["macaddress"]
-            print("{1:<{0}}{2:<{3}}{4}".format(
-                padding, "", ip if ip else "<not set>", len_ip,
-                mac if mac else "<not set>"))
-
-    # print aaaa records
-    if aaaa_records:
-        print("{1:<{0}}{2:<{3}}{4}".format(padding, "AAAA_Records:", "IP", len_ip, "MAC"))
-        for record in aaaa_records:
-            ip = record["ipaddress"]
-            mac = record["macaddress"]
-            print("{1:<{0}}{2:<{3}}{4}".format(
-                padding, "", ip if ip else "<not set>", len_ip,
-                mac if mac else "<not set>"))
-
-
-def print_ttl(ttl: int, padding: int = 14) -> None:
-    """Pretty print given ttl"""
-    assert isinstance(ttl, int) or ttl is None
-    print("{1:<{0}}{2}".format(padding, "TTL:", ttl or "(Default)"))
-
-
-def print_hinfo(hid: str, padding: int = 14) -> None:
-    """Pretty given hinfo id"""
-    hinfos = hinfo_dict()
-    hid = str(hid)
-    hinfo = hinfos[hid]
-    print("{1:<{0}}cpu={2} os={3}".format(padding, "Hinfo:", hinfo[0], hinfo[1]))
-
-
-def print_hinfo_list(hinfos: HinfoDict, padding: int = 14) -> None:
-    """Pretty print a dict of host infos"""
-    if len(hinfos) == 0:
-        print("No hinfo presets.")
-        return
-    max_len = max([len(x[0]) for x in hinfos.values()])
-    print("{1:<{0}}    {2:<{3}} {4}".format(padding, "Id", "CPU", max_len, "OS"))
-    for hid in sorted(hinfos.keys()):
-        hinfo = hinfos[hid]
-        print(
-            "{1:<{0}} -> {2:<{3}} {4}".format(padding, hid, hinfo[0], max_len, hinfo[1]))
-
-
-def print_srv(srv: dict, padding: int = 14) -> None:
-    """Pretty print given srv"""
-    print("{1:<{0}} SRV {2:^6} {3:^6} {4:^6} {5}".format(
-        padding,
-        srv["name"],
-        srv["priority"],
-        srv["weight"],
-        srv["port"],
-        srv["target"],
-    ))
-
-
-def print_loc(loc: str, padding: int = 14) -> None:
-    """Pretty print given loc"""
-    if loc is None:
-        return
-    assert isinstance(loc, str)
-    print("{1:<{0}}{2}".format(padding, "Loc:", loc))
-
-
-def print_cname(cname: str, host: str, padding: int = 14) -> None:
-    """Pretty print given cname"""
-    print("{1:<{0}}{2} -> {3}".format(padding, "Cname:", cname, host))
-
-
-def print_txt(txt: str, padding: int = 14) -> None:
-    """Pretty print given txt"""
-    if txt is None:
-        return
-    assert isinstance(txt, str)
-    print("{1:<{0}}{2}".format(padding, "TXT:", txt))
-
-
-def print_naptr(naptr: dict, host_name: str, padding: int = 14) -> None:
-    """Pretty print given txt"""
-    assert isinstance(naptr, dict)
-    assert isinstance(host_name, str)
-    print("{1:<{0}} NAPTR {2} {3} \"{4}\" \"{5}\" \"{6}\" {7}".format(
-        padding,
-        host_name,
-        naptr["preference"],
-        naptr["order"],
-        naptr["flag"],
-        naptr["service"],
-        naptr["regex"] or "",
-        naptr["replacement"],
-    ))
-
-
-def print_ptr(ip: str, host_name: str, padding: int = 14) -> None:
-    """Pretty print given txt"""
-    assert isinstance(ip, str)
-    assert isinstance(host_name, str)
-    print("{1:<{0}} PTR {2}".format(padding, ip, host_name))
-
-
-################################################################################
-#                                                                              #
 #   Validation functions                                                       #
 #                                                                              #
 ################################################################################
@@ -616,11 +431,6 @@ def is_valid_email(email: typing.AnyStr) -> bool:
         except ValueError:
             return False
     return True if re.match(r"^[^\s@]+@[^\s@]+\.[^\s@]+$", email) else False
-
-
-def is_valid_loc(loc: str) -> bool:
-    # TODO LOC: implement validate loc
-    return True
 
 
 def is_valid_location_tag(loc: str) -> bool:
